@SW-17: updated helper and added minor changes
