+++ SmartWork/SmartWork.Configuration/DBSettings.cs
@@ -1,25 +1,15 @@
 ﻿using Microsoft.Extensions.Configuration;
-using System.IO;
-using System.Text.RegularExpressions;
 
 namespace SmartWork.Configuration
 {
     public class DBSettings
     {
+        const string PATH_TO_PASSWORD = @"C:\Users\Dmitriy\Desktop\Workspace\Univercity\SecureData\dbConnectionData.txt";
         public static string GetDBConnectionString()
         {
-            var configuretion = ConfigurationSettings.GetConfiguration();
-            var dataBaseConnectionString = configuretion.GetConnectionString("DataBaseConnection");
-            var path = @"C:\Users\Dmitriy\Documents\SmartWork\dbConnectionData.txt";
-            var dbPassword = string.Empty;
-            using (var reader = new StreamReader(path))
-            {
-                dbPassword = reader.ReadToEnd();
-                var regex = new Regex(@"^password=\[.+\]");
-                var mathValue = regex.Match(dbPassword).Value;
-                dbPassword = dbPassword[(dbPassword.IndexOf('[') + 1)..];
-                dbPassword = dbPassword.Remove(dbPassword.LastIndexOf(']'));
-            }
+            var configuration = ConfigurationSettings.GetConfiguration();
+            var dataBaseConnectionString = configuration.GetConnectionString("DataBaseConnection");
+            Helper.GetSecurePasswordForDb(PATH_TO_PASSWORD, out string dbPassword);
             dataBaseConnectionString = dataBaseConnectionString.Replace("********", dbPassword);
             return dataBaseConnectionString;
         }
